Move export code to own module
