+++ corehq/blobs/management/commands/run_blob_export.py
@@ -4,7 +4,7 @@ import logging
 import os
 import sys
 from django.core.management import BaseCommand, CommandError
-from corehq.blobs.migrate import EXPORTERS
+from corehq.blobs.export import EXPORTERS
 from corehq.blobs.zipdb import get_export_filename
 from corehq.util.decorators import change_log_level
